processing/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
piexif>=1.1.3
requests>=2.28.0
